UTIL/mpmapas_python_utils/requirements.txt
# -*- coding: utf-8 -*-
cchardet>=2.1.7
chardet>=4.0.0
numpy>=1.19.5
openpyxl>=3.0.6
pandas>=1.2.1
pathlib>=1.0.1
psycopg2-binary>=2.8.6
pyjavaproperties>=0.7
PyYAML>=5.4.1
SQLAlchemy>=1.4.2
urllib3>=1.26.3
xlrd>=2.0.1
